+++ lib/views/home_view.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class HomeView extends StatelessWidget {
+  const HomeView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text("PR 연습하기"),
+      ),
+      body: Center(
+        child: Column(
+          children: [
+            ElevatedButton(
+              onPressed: () {},
+              child: const Text("이름"),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
